Reorganize exceptions into appropriate namespaces.
