+++ wrappers/dotnet/indy-sdk-dotnet-test/AnonCredsTests/ProverCreateProofTest.cs
@@ -1,4 +1,5 @@
 ﻿using Hyperledger.Indy.AnonCredsApi;
+using Hyperledger.Indy.WalletApi;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json.Linq;
 using System;
