+++ wrappers/dotnet/indy-sdk-dotnet-test/SignusTests/ReplaceKeysApplyTest.cs
@@ -1,4 +1,5 @@
 ﻿using Hyperledger.Indy.SignusApi;
+using Hyperledger.Indy.WalletApi;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
