+++ wrappers/dotnet/indy-sdk-dotnet/AnonCredsApi/AnoncredsAccumulatorFullException.cs
@@ -1,0 +1,16 @@
+﻿namespace Hyperledger.Indy.AnonCredsApi
+{
+    /// <summary>
+    /// Exception thrown when an anoncreds accululator is full.
+    /// </summary>
+    public class AnoncredsAccumulatorFullException : IndyException
+    {
+        const string message = "The anoncreds accumulator is full.";
+
+        internal AnoncredsAccumulatorFullException(int sdkErrorCode) : base(message, sdkErrorCode)
+        {
+
+        }
+    }
+
+}
