+++ wrappers/dotnet/indy-sdk-dotnet/AnonCredsApi/ClaimRevokedException.cs
@@ -1,0 +1,16 @@
+﻿namespace Hyperledger.Indy.AnonCredsApi
+{
+    /// <summary>
+    /// Exception thrown when a claim has been revoked.
+    /// </summary>
+    public class ClaimRevokedException : IndyException
+    {
+        const string message = "The claim has been revoked.";
+
+        internal ClaimRevokedException(int sdkErrorCode) : base(message, sdkErrorCode)
+        {
+
+        }
+    }
+
+}
