+++ wrappers/dotnet/indy-sdk-dotnet/AnonCredsApi/DuplicateMasterSecretNameException.cs
@@ -1,0 +1,16 @@
+﻿namespace Hyperledger.Indy.AnonCredsApi
+{
+    /// <summary>
+    /// Exception thrown when an attempt to create a master-secret with the same name as an existing master-secret.
+    /// </summary>
+    public class DuplicateMasterSecretNameException : IndyException
+    {
+        const string message = "Another master-secret with the specified name already exists.";
+
+        internal DuplicateMasterSecretNameException(int sdkErrorCode) : base(message, sdkErrorCode)
+        {
+
+        }
+    }
+
+}
