+++ wrappers/dotnet/indy-sdk-dotnet/PoolApi/PoolLedgerTerminatedException.cs
@@ -1,0 +1,16 @@
+﻿namespace Hyperledger.Indy.PoolApi
+{
+    /// <summary>
+    /// Exception thrown when the pool ledger was terminated.
+    /// </summary>
+    public class PoolLedgerTerminatedException : IndyException
+    {
+        const string message = "The pool ledger was terminated.";
+
+        internal PoolLedgerTerminatedException(int sdkErrorCode) : base(message, sdkErrorCode)
+        {
+
+        }
+    }
+
+}
